+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/userInterface/panel/NotificationType.kt
@@ -1,7 +1,7 @@
 package com.github.naoyukik.intellijplugingithubnotifications.userInterface.panel
 
 enum class NotificationType(val displayName: String) {
-    Unselected(""),
+    UNSELECTED("<Choose type>"),
     PULL_REQUEST_OPEN("Pull Request Open"),
     PULL_REQUEST_MERGED("Pull Request Merged"),
     PULL_REQUEST_CLOSED("Pull Request Closed"),