+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/userInterface/panel/NotificationUnread.kt
@@ -1,0 +1,7 @@
+package com.github.naoyukik.intellijplugingithubnotifications.userInterface.panel
+
+enum class NotificationUnread(val displayName: String) {
+    ALL("Read and Unread"),
+    UNREAD("Unread only"),
+    READ("Read only"),
+}