+++ CHANGELOG.md
@@ -6,6 +6,34 @@
 
 ### Added
 
+- Add Read filter
+
+### Fixed
+
+- Set a minimum value for column height
+
+## [0.10.0] - 2025-03-16
+
+### Added
+
+- Reset the latest fetch time on settings apply
+- Implement label column
+- Implement label and unread filter
+
+### Fixed
+
+- Optimize type filter
+
+## [0.9.0] - 2025-03-12
+
+### Added
+
+- Implement Reviewers filter
+
+## [0.8.0] - 2025-03-10
+
+### Added
+
 - Implement Type filter
 
 ## [0.7.0] - 2025-02-21
@@ -94,7 +122,10 @@ Initial release of the GitHub Notifications IntelliJ Plugin.
 
 Done is better than perfect.
 
-[Unreleased]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.8.0...v0.9.0
+[0.8.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.4.3...v0.5.0